requirements.txt
# api
fastapi[standard]
pydantic
python-dotenv

# ml
google-generativeai
groq

# vector db
langchain_community
langchain_text_splitters
chromadb
pypdf

# for notebooks and testing
ipykernel
pip